refactor: add createVariableRule() helper
